Update parse.js
Parsing means analyzing and converting a program into an internal format that a runtime environment can actually run, for example the JavaScript engine inside browsers. The browser parses HTML into a DOM tree. JS compiler parses a string and returns a JavaScript object. The string has to be written in JSON format. The JSON.parse() method can optionally transform the ...A parser basically reads our code line by line and checks it for errors and stops execution if any, if the code is free of syntactic errors,